+++ lib/widgets/post/post_title.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import '../../url_launcher.dart';
 import '../../util/observer_consumers.dart';
 import '../cached_network_image.dart';
+import '../fullscreenable_image.dart';
 import 'post_store.dart';
 
 class PostTitle extends StatelessWidget {
@@ -61,6 +62,20 @@ class PostTitle extends StatelessWidget {
                   ),
                 ),
               ],
+              if (store.hasMedia && url != null) ...[
+                FullscreenableImage(
+                  url: url,
+                  child: CachedNetworkImage(
+                    width: 70,
+                    height: 70,
+                    imageUrl: url,
+                    errorBuilder: (_, ___) => const Icon(Icons.warning),
+                    loadingBuilder: (context, progress) =>
+                        CircularProgressIndicator.adaptive(
+                            value: progress?.progress),
+                  ),
+                )
+              ]
             ],
           ),
         );